refactor(settings): extract re-auth gate into useReauthVerification
Move the promise-based verification modal state machine verbatim into
_hooks/use-reauth-verification.ts: verifyModal/code/error/loading state,
requestVerification (await-able), closeVerifyModal, handleVerifySubmit,
and the 6-digit auto-submit effect. Owns its own Settings i18n; takes
mfaEnabled. Pruned now-unused useCallback import. Behavior-identical;
page.tsx 1522->1458 LOC.
